# core-common-distributor-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-distributor-mgmt
  description: Distributor Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/distributors
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/lending-contracts/{id}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contract by ID
      description: Returns a lending contract based on its ID
      operationId: getLendingContractById
      parameters:
        - name: id
          in: path
          description: ID of the lending contract to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Leasing contract not found
        '500':
          description: Internal server error
    put:
      tags:
        - Lending Contract
      summary: Update an existing lending contract
      description: Updates an existing lending contract with the provided information
      operationId: updateLendingContract
      parameters:
        - name: id
          in: path
          description: ID of the lending contract to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LendingContractDTO'
        required: true
      responses:
        '200':
          description: Leasing contract successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '400':
          description: Invalid lending contract data provided
        '404':
          description: Leasing contract not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Lending Contract
      summary: Delete a lending contract
      description: Deletes a lending contract based on its ID
      operationId: deleteLendingContract
      parameters:
        - name: id
          in: path
          description: ID of the lending contract to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Leasing contract successfully deleted
        '404':
          description: Leasing contract not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/contracts/{contractId}:
    get:
      tags:
        - Distributor Contract
      summary: Get contract by ID
      description: Retrieve a specific contract associated with a distributor by its unique identifier
      operationId: getById_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: contractId
          in: path
          description: Unique identifier of the contract
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Contract found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorContractDTO'
        '404':
          description: Contract or distributor not found
    put:
      tags:
        - Distributor Contract
      summary: Update contract
      description: Update an existing contract associated with a distributor
      operationId: update_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: contractId
          in: path
          description: Unique identifier of the contract
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorContractDTO'
        required: true
      responses:
        '200':
          description: Contract successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorContractDTO'
        '400':
          description: Invalid contract data
        '404':
          description: Contract or distributor not found
    delete:
      tags:
        - Distributor Contract
      summary: Delete contract
      description: Delete a contract associated with a distributor from the system
      operationId: delete_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: contractId
          in: path
          description: Unique identifier of the contract
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Contract successfully deleted
        '404':
          description: Contract or distributor not found
...
  /api/v1/lending-contracts/{id}/approve:
    post:
      tags:
        - Lending Contract
      summary: Approve a lending contract
      description: Approves a lending contract and updates its status
      operationId: approveLendingContract
      parameters:
        - name: id
          in: path
          description: ID of the lending contract to approve
          required: true
          schema:
            type: string
            format: uuid
        - name: approvedBy
          in: query
          description: ID of the user approving the contract
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Leasing contract successfully approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '400':
          description: Invalid approval request
        '404':
          description: Leasing contract not found
        '500':
          description: Internal server error
  /api/v1/lending-contracts/filter:
    post:
      tags:
        - Lending Contract
      summary: Filter lending contracts
      description: Returns a paginated list of lending contracts based on filter criteria
      operationId: filterLendingContracts
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestLendingContractDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json: {}
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
  /api/v1/lending-contracts:
    post:
      tags:
        - Lending Contract
      summary: Create a new lending contract
      description: Creates a new lending contract with the provided information
      operationId: createLendingContract
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LendingContractDTO'
        required: true
      responses:
        '201':
          description: Leasing contract successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '400':
          description: Invalid lending contract data provided
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/contracts/filter:
    post:
      tags:
        - Distributor Contract
      summary: Filter contracts for a distributor
      description: Retrieve a paginated list of contracts associated with a specific distributor based on filtering criteria
      operationId: filter_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorContractDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered contracts
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Distributor not found
...
  /api/v1/lending-contracts/status/{status}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contracts by status
      description: Returns all lending contracts with a specific status
      operationId: getLendingContractsByStatus
      parameters:
        - name: status
          in: path
          description: Status value to filter lending contracts
          required: true
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '400':
          description: Invalid status value
        '500':
          description: Internal server error
  /api/v1/lending-contracts/product/{productId}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contracts by product
      description: Returns all lending contracts associated with a specific product
      operationId: getLendingContractsByProductId
      parameters:
        - name: productId
          in: path
          description: ID of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Product not found
        '500':
          description: Internal server error
  /api/v1/lending-contracts/party/{partyId}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contracts by party
      description: Returns all lending contracts associated with a specific party
      operationId: getLendingContractsByPartyId
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Party not found
        '500':
          description: Internal server error
  /api/v1/lending-contracts/distributor/{distributorId}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contracts by distributor
      description: Returns all lending contracts associated with a specific distributor
      operationId: getLendingContractsByDistributorId
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    FilterRequestLendingContractDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/LendingContractDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestDistributorContractDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorContractDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for contracts
...